Add a notice at the bottom of move embeds
If an image link is set in the embed but is no longer valid, Discord
will not show anything as if a link was never set. Now, when an
attachment is set, regardless of if it is visible, a notice will be at
the bottom of the embed making sure people are aware if any errors
with the link occur.